+++ LookupAnything/Framework/Lookups/Items/ItemSubject.cs
@@ -691,19 +691,30 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
 
             // get community center
             CommunityCenter communityCenter = Game1.locations.OfType<CommunityCenter>().First();
-            if (communityCenter.areAllAreasComplete() && communityCenter.isBundleComplete(36))
-                yield break;
+            bool IsBundleOpen(int id)
+            {
+                try
+                {
+                    return !communityCenter.isBundleComplete(id);
+                }
+                catch
+                {
+                    return false; // invalid bundle data
+                }
+            }
 
             // get bundles
-            foreach (BundleModel bundle in this.GameHelper.GetBundleData())
+            if (!communityCenter.areAllAreasComplete() || IsBundleOpen(36))
             {
-                // ignore completed bundle
-                if (communityCenter.isBundleComplete(bundle.ID))
-                    continue;
+                foreach (BundleModel bundle in this.GameHelper.GetBundleData())
+                {
+                    if (!IsBundleOpen(bundle.ID))
+                        continue;
 
-                bool isMissing = this.GetIngredientsFromBundle(bundle, item).Any(p => this.IsIngredientNeeded(bundle, p));
-                if (isMissing)
-                    yield return bundle;
+                    bool isMissing = this.GetIngredientsFromBundle(bundle, item).Any(p => this.IsIngredientNeeded(bundle, p));
+                    if (isMissing)
+                        yield return bundle;
+                }
             }
         }
 
@@ -800,7 +811,7 @@ namespace Pathoschild.Stardew.LookupAnything.Framework.Lookups.Items
             // present in the community center data. This seems to be caused by some mods like
             // Challenging Community Center Bundles in some cases.
             if (!communityCenter.bundles.TryGetValue(bundle.ID, out bool[] items) || ingredient.Index >= items.Length)
-                return true; 
+                return true;
 
             return !items[ingredient.Index];
         }